+++ HashTable.h
@@ -17,16 +17,16 @@ class Employee;
 
 class HashTable {
     DynamicArray<AVLTree<Employee, CompareEmpByID>> array;
-    int arraySize;
-    int numberOfElements;
+    long arraySize;
+    long numberOfElements;
 public:
     HashTable() : array(), arraySize(INIT_SIZE), numberOfElements(0) {}
     ~HashTable() = default;
 
-    int getSize() const {
+    long getSize() const {
         return numberOfElements;
     }
-    unsigned int hash(shared_ptr<Employee> emp) {
+    unsigned long hash(shared_ptr<Employee> emp) {
         return emp->id % arraySize;
     }
 
@@ -47,7 +47,7 @@ public:
 //        }
     }
 
-    bool find(int id) {
+    bool find(long id) {
         shared_ptr<Employee> toFind = shared_ptr<Employee>(new Employee(id, 0, 0, nullptr));
 
         return array[hash(toFind)].find(toFind) != nullptr;
@@ -63,10 +63,10 @@ public:
 
     void rehash() {
         DynamicArray<AVLTree<Employee, CompareEmpByID>> newArray(arraySize * 2);
-        int oldSize = arraySize;
+        long oldSize = arraySize;
         arraySize = 2 * arraySize;
         try{
-            for (int i = 0; i < oldSize; ++i) {
+            for (long i = 0; i < oldSize; ++i) {
                 for (AVLTree<Employee, CompareEmpByID>::AVLIter it = array[i].begin(); it != array[i].end() ; ++it){
                     newArray[hash(*it)].insert(*it);
                 }
@@ -82,10 +82,10 @@ public:
 
     void rehashDown() {
         DynamicArray<AVLTree<Employee, CompareEmpByID>> newArray(arraySize / 2);
-        int oldSize = arraySize;
+        long oldSize = arraySize;
         arraySize = arraySize / 2;
         try{
-            for (int i = 0; i < oldSize; ++i) {
+            for (long i = 0; i < oldSize; ++i) {
                 for (AVLTree<Employee, CompareEmpByID>::AVLIter it = array[i].begin(); it != array[i].end() ; ++it){
                     newArray[hash(*it)].insert(*it);
                 }
@@ -102,16 +102,16 @@ public:
 
     void merge(const HashTable& toMergeFrom) {
         DynamicArray<AVLTree<Employee, CompareEmpByID>> newArray((arraySize + toMergeFrom.arraySize) * 2);
-        int oldSize = arraySize;
-        int toMergeSize = toMergeFrom.arraySize;
+        long oldSize = arraySize;
+        long toMergeSize = toMergeFrom.arraySize;
         arraySize = (arraySize + toMergeFrom.arraySize) * 2;
         try{
-            for (int i = 0; i < oldSize; ++i) {
+            for (long i = 0; i < oldSize; ++i) {
                 for (AVLTree<Employee, CompareEmpByID>::AVLIter it = array[i].begin(); it != array[i].end() ; ++it){
                     newArray[hash(*it)].insert(*it);
                 }
             }
-            for (int i = 0; i < toMergeSize; ++i) {
+            for (long i = 0; i < toMergeSize; ++i) {
                 for (AVLTree<Employee, CompareEmpByID>::AVLIter it = toMergeFrom.array[i].begin(); it != toMergeFrom.array[i].end() ; ++it){
                     newArray[hash(*it)].insert(*it);
                 }
